requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pkg-resources==0.0.0
python-consul==1.1.0
PyYAML==5.3
redis==3.4.1
requests==2.23.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
